requirements.txt
appdirs==1.4.0
chardet==2.3.0
cssselect==1.0.0
cssutils==1.0.1
diff-match-patch==20121119
dj-database-url==0.4.1
Django==1.10.3
django-anymail==0.8
django-appconf==1.0.2
django-bootstrap3==8.1.0
django-datatable==0.2.1
django-premailer==0.2.0
docker==2.1.0
docker-pycreds==0.2.1
gunicorn==19.6.0
lxml==3.7.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
premailer==3.0.0
psycopg2==2.6.2
pycrypto==2.6.1
pyparsing==2.1.10
pytz==2016.10
rcssmin==1.0.6
reportlab==3.3.0
requests==2.13.0
rjsmin==1.0.12
simplejson==3.10.0
six==1.10.0
websocket-client==0.40.0
